Fix Detekt issues
